+++ project/v2/shop/app/shop_srv/goods/srv/internal/data/brand.go
@@ -7,9 +7,9 @@ import (
 )
 
 type BrandsStore interface {
-	Get(ctx context.Context, ID int32) (*do.BrandsDO, error)
+	Get(ctx context.Context, ID int64) (*do.BrandsDO, error)
 	List(ctx context.Context, opts metav1.ListMeta, orderby []string) (*do.BrandsDOList, error)
 	Create(ctx context.Context, brands *do.BrandsDO) error
 	Update(ctx context.Context, brands *do.BrandsDO) error
-	Delete(ctx context.Context, ID uint64) error
+	Delete(ctx context.Context, ID int64) error
 }
